+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # pytest-djangoapp changelog
 
 
-### Unreleased
+### v1.7.0 [2025-12-13]
 * ++ Admin client. Added '.call_add()'.
 * ++ Admin client. Added support for data update in '.call_change()'.
 * ++ Admin client. Exposed 'url_listing' and 'url_add' attributes.